ex9 python2

+++ examples/ex1/run_ex1.sh
@@ -9,11 +9,11 @@
 # ====================================================
 # user settings
 
-EXEDLPOLY=/home/filipe/build/dl_poly/execute/DLPOLY.X
-EXEMDFF=mdff20.x
+EXEDLPOLY=/home/filipe/dev/dl_poly/execute/DLPOLY.X
+EXEMDFF=/home/filipe/dev/mdff20/src/mdff20.x
 
+mdff=true
 do_dlpoly=true
-
 # =====================================================
 
 sep="=================================================="
@@ -22,17 +22,19 @@ echo "#Example 1: LJ (argon) fcc structure at low temperature"
 echo "The configuration is readed in POSFF"
 echo "more info in control.s and stdout"
 
-rm -rf mdff/
-mkdir mdff
-cd mdff
-echo 
-echo "# mdff calculation "
-cp ../config/control.s .
-cp ../config/POSFF .
-$EXEMDFF control.s > stdout
-#poszi.py -i OSZIFF -n 
-grep e+ OSZIFF | sed -n 'p;n' | awk '{print $2,$3}' > ene
-cd ..
+if $mdff; then
+    rm -rf mdff/
+    mkdir mdff
+    cd mdff
+    echo 
+    echo "# mdff calculation requested"
+    cp ../config/control.s .
+    cp ../config/POSFF .
+    $EXEMDFF control.s > stdout
+    #poszi.py -i OSZIFF -n 
+    grep e+ OSZIFF | sed -n 'p;n' | awk '{print $2,$3}' > ene
+    cd ..
+fi
 
 if $do_dlpoly; then
         echo "# dlpoly calculation requested "
@@ -43,12 +45,12 @@ if $do_dlpoly; then
         cp ../config/FIELD .
         cp ../config/CONFIG .
         $EXEDLPOLY
-	read_statis > ene
+	/home/filipe/dev/mdff20/bin/read_statis > ene
         cd ..
 fi
 
 lr=$(grep "long range correction (energy)" mdff/stdout |  awk '{print $NF}')
-echo $lr
+echo "long range correction $lr"
 cat > plot << eof
 #!/usr/bin/gnuplot -persist 
 set term x11
